refactor(imagetool.manager): refactor console and jitted functions to private modules to improve startup time
